+++ packages/terra-toggle/src/Toggle.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import AnimateHeight from 'react-animate-height';
 import 'terra-base/lib/baseStyles';
-import styles from './Toggle.scss';
+import styles from './Toggle.module.scss';
 
 const cx = classNames.bind(styles);