+++ packages/terra-time-input/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 Unreleased
 ----------
 ### Changed
+* Add .module extension to SCSS file(s)
 * Replace legacy Input with the new terra-form-input package
 * Give desktop time input the same styles as terra-form-input.